backend/requirements.txt
asgiref==3.8.1
Django==3.2.6
djangorestframework==3.12.4
psycopg2-binary==2.9.1
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2
django-cors-headers
